make myci-deb-prepare.sh substitute soname var in filenames

+++ src/myci-deb-prepare.sh
@@ -3,6 +3,8 @@
 # we want exit immediately if any command fails and we want error in piped commands to be preserved
 set -eo pipefail
 
+script_dir="$(dirname $0)/"
+
 # this script is used for preparing the debian package before building with dpkg-buildpackage.
 
 while [[ $# > 0 ]] ; do
@@ -52,7 +54,15 @@ listOfInstalls=$(ls $debianization_dir/*.install.in 2>/dev/null || true) # allow
 
 for i in $listOfInstalls; do
 	echo "applying soname to $i"
-	sed -e "s/\$(soname)/$soname/g" $i > ${i%.install.in}$soname.install
+
+	# BACKWARDS COMPATIBILITY: in case file name does not contain $ sign, then do not
+	#                          substitute $(soname) variable in file name, but just append soname value to it.
+	#                          Otherwise do $(soname) substitution in file name.
+	if [ $(echo "$i" | sed -n -e "s/.*\$.*/true/p") == "true" ]; then
+		${script_dir}myci-subst-var.sh --var soname --val $soname $i
+	else
+		sed -e "s/\$(soname)/$soname/g" $i > ${i%.install.in}$soname.install
+	fi
 done
 
 echo "applying soname to $debianization_dir/control.in"
